+++ frontend/js/src/brainzplayer/BrainzPlayerUI.tsx
@@ -12,13 +12,13 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { IconDefinition } from "@fortawesome/fontawesome-common-types"; // eslint-disable-line import/no-unresolved
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { isNaN as _isNaN } from "lodash";
-import { ToastContainer, toast } from "react-toastify";
-import { ToastMsg } from "./BrainzPlayer";
+import { toast } from "react-toastify";
 import ProgressBar from "./ProgressBar";
 import GlobalAppContext from "../utils/GlobalAppContext";
 import MenuOptions from "./MenuOptions";
 import { millisecondsToStr } from "../playlists/utils";
 import { getRecordingMBID, getRecordingMSID } from "../utils/utils";
+import { ToastMsg } from "../notifications/Notifications";
 
 type BrainzPlayerUIProps = {
   currentDataSourceName?: string;
