+++ frontend/js/src/notifications/Notifications.tsx
@@ -1,3 +1,19 @@
+import React from "react";
+
+export type ToastMsgProps = {
+  title: string;
+  message: string | JSX.Element;
+};
+
+export function ToastMsg({ title, message }: ToastMsgProps) {
+  return (
+    <div>
+      <b>{title}</b>
+      <p>{message}</p>
+    </div>
+  );
+}
+
 /**
  * Older implementations and Safari use the deprecated callback signature
  * of Notification.requestPermission instead of the Promise-based signature
